backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.20",
    "langchain-google-genai>=2.0.10",
    "langchain-community>=0.1.0",
    "pydantic>=2.10.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "aiosqlite>=0.22.1",
    "langchain-tavily>=0.2.18",
    "langfuse>=4.5.1",
    "pydantic-settings>=2.13.1",
    "rich>=14.3.3",
    "fastmcp>=0.4.1",
]
[tool.fastapi]
entrypoint = "app.main:app"

[tool.ruff]
# Configuração base do Ruff
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# Habilita regras recomendadas
select = ["E", "F", "I", "N", "UP", "B"]
ignore = []

[tool.ruff.format]
# Configurações de formatação (estilo Black)
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
